Refactor vision_constants and update GPSR and HRIC commands to use area_check for filtering detections. Removed obsolete filter_class utility to streamline detection logic and improve maintainability.
