+++ frida_constants/frida_constants/vision_constants.py
@@ -125,8 +125,6 @@ COUNT_BY_PERSON_TOPIC = "/vision/gpsr/count_by_person"
 COUNT_BY_OBJECTS_TOPIC = "/vision/gpsr/count_by_objects"
 COUNT_BY_POSE_TOPIC = "/vision/gpsr/count_by_pose"
 COUNT_BY_GESTURE_TOPIC = "/vision/gpsr/count_by_gesture"
-FILTER_CLASS_TOPIC = "/vision/gpsr/filter_class"
-FILTER_CLASS_ROOM_TOPIC = "/vision/gpsr/filter_class_room"
 IMAGE_TOPIC = "/vision/gpsr/img_detection"
 POSE_GESTURE_TOPIC = "/vision/gpsr/pose_gesture_detection"
 READ_QR_TOPIC = "/vision/gpsr/read_qr"
